A whole lot of changes, specifically around handling issues around lack of memory and not processing repos/files that have not changed.
Now working to track the last CommitID for a repo or file, and will only process if this has changed.

In some cases for the very large repos, we get an out of memory error when attempting to download the repo. In this case I am just flagging the files as part of an "out of memory" error and proceeding. Later on I will look at other ways to stream this into a file directly instead of loading into memory.

Work in progress, still more to do...