Had to modify when parsing package references in .Net projects: - make case-insensitive when getting the version - version could be either an atttribute of the PackageReference, or it could be a child node..
Also modified the Saving of Package Refs, as there are cases where multiple references to the same package with different versions is possible. I imagine the compiler handles this, but was breaking the code to detect and update existing data in the database.  Now I just get the latest version of any one package by version.