Cloud Migration and package cleanup (#17)
* Bug cropped up, not sure how I missed this. Needed to make buildProperties static (and made it readonly as well for now).

* Clearing up Mend security issues more directly.

* For all of the "secrets", I have created a separate class (ConfidentialSettings) with a static class of Values. This contains the set of secrets that are used elsewhere in the various TFM components.

Right now this class reads the values from environment variables, that I presume can be loaded from the Key Vault(s) for TFM. However it can be easily amended if we need to read these values elsewhere.

* Changing the environment names to match the production names in KV

* Had to modify when parsing package references in .Net projects:
- make case-insensitive when getting the version
- version could be either an atttribute of the PackageReference, or it could be a child node..

Also modified the Saving of Package Refs, as there are cases where multiple references to the same package with different versions is possible. I imagine the compiler handles this, but was breaking the code to detect and update existing data in the database.  Now I just get the latest version of any one package by version.

---------

Co-authored-by: Andreas Mertens <Andreas Mertens Andreas.Mertens@worksafebc.com>