Adding code to extract more of the versions from packages in the .Net project files.
NOTE: there is still a much older format that I need to address for NuGet packages, but this is s a good start. I want next add the command line parameters for project and repo so I can refresh a single repo. We want this anyways so that we can trigger of a commit to do a scan of just the repo changes.  However this also allows me to debug with out requiring a full scan to be performed, speeding up debugging and validation.