Cloud Migration, part 1 (#18)
* Bug cropped up, not sure how I missed this. Needed to make buildProperties static (and made it readonly as well for now).

* Clearing up Mend security issues more directly.

* For all of the "secrets", I have created a separate class (ConfidentialSettings) with a static class of Values. This contains the set of secrets that are used elsewhere in the various TFM components.

Right now this class reads the values from environment variables, that I presume can be loaded from the Key Vault(s) for TFM. However it can be easily amended if we need to read these values elsewhere.

* Changing the environment names to match the production names in KV

* Had to modify when parsing package references in .Net projects:
- make case-insensitive when getting the version
- version could be either an atttribute of the PackageReference, or it could be a child node..

Also modified the Saving of Package Refs, as there are cases where multiple references to the same package with different versions is possible. I imagine the compiler handles this, but was breaking the code to detect and update existing data in the database.  Now I just get the latest version of any one package by version.

* Adding debug information to diagnose some configuration loading issues

* Removing console code to read key at end of processing

* Replacing debug writeline statements with console writeline. This can then be displayed in the run of the pipeline

* Need to confirm what the checkout directory is.

Also setting forceDetails to true, to ensure we are actually running on all files

* Adding a lot of debugging comments to display in the Console to determine why the Files table in the database is not being populated.

* Having issue when loading a specific configuration file. Adding more console status messages to track this.

* More debug info in configuration code

* Temporarily disabling the RepoScan and FileScan, only allow the FileDetails Scan to run.

* Refactoring code to eliminate the need for two db connections when reading file details

* Set up almost all of the Console.Writeline commands to be inside an #if DEBUG block, so that they do not run in a Release build.  Otherwise the Run pipeline can lose output which might be required if an error occurs.

Also re-enabled the first two scan passes.

* Making the thread start/end writelines only work in DEBUG builds

* reset forceDetails to false. For got to do ths in the last commit.

* Adding code to extract more of the versions from packages in the .Net project files.

NOTE: there is still a much older format that I need to address for NuGet packages, but this is s a good start. I want next add the command line parameters for project and repo so I can refresh a single repo. We want this anyways so that we can trigger of a commit to do a scan of just the repo changes.  However this also allows me to debug with out requiring a full scan to be performed, speeding up debugging and validation.

---------

Co-authored-by: Andreas Mertens <Andreas Mertens Andreas.Mertens@worksafebc.com>