Created a new parameter -x that takes a comma-seperated list of project IDs (GUIDs) that are excluded from scans. This helps when running a full refresh in the pipeline to break up the scan into multiple scans so that it doesn't time out the pipeline run.
Fixed an issue when checking out a repo - sometimes the target folder is not found. Force it to always check and create just before checkout.

Fixed an issue that happens when a repo is moved from one project to another. The repo has the same GUID Id. So we end up with two instances in the TFM data - one deleted in original project, and another that is active.  Some of my queries were not checking the deleted flag and was expecting a single instance.